fix: CustomerSubscriptionResponse removal of unnecessary property

+++ GestaoPortfolioInvestimento/GestaoPortfolioInvestimento/Application/DTOs/CustomerSubscription/CustomerSubscriptionCreateRequest.cs
@@ -14,6 +14,6 @@ namespace InvestmentPortfolioManagement.Application.DTOs.CustomerSubscription
         [Range(1, int.MaxValue, ErrorMessage = "Quantity must be greater than zero.")]
         public int Quantity { get; set; }
 
-        public Double TotalValue { get; set; }
+        public decimal TotalValue { get; set; }
     }
 }
